backend/requirements.txt
Django==4.2.7
django-comment-migrate==0.1.7
django-cors-headers==4.3.0
django-filter==23.3
django-ranged-response==0.2.0
djangorestframework==3.14.0
django-restql==0.15.3
django-simple-captcha==0.5.20
django-timezone-field==6.0.1
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
mysqlclient==2.2.0
pypinyin==0.49.0
ua-parser==0.18.0
pyparsing==3.1.1
openpyxl==3.1.2
requests==2.31.0
typing-extensions==4.8.0
tzlocal==5.1
channels==3.0.5
channels-redis==4.1.0
websockets==11.0.3
user-agents==2.2.0
six==1.16.0
whitenoise==6.6.0
psycopg2==2.9.9
uvicorn==0.23.2
gunicorn==21.2.0
gevent
Pillow==10.1.0
dvadmin-celery==1.0.5
django-extensions
numpy==1.24.4